@@ -10,17 +10,17 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=7.3",
         "composer-plugin-api": "^1.0 || ^2.0"
     },
     "require-dev": {
-        "composer/composer": "^1.2",
-        "phpunit/phpunit": ">=4.8.35 <6.0",
-        "codeclimate/php-test-reporter": "dev-master"
+        "composer/composer": "^2.0",
+        "phpunit/phpunit": "^8.5",
+        "phpmd/phpmd": "^2.9"
     },
     "autoload": {
-        "psr-0": {
-            "NodejsPhpFallback\\": "src/"
+        "psr-4": {
+            "NodejsPhpFallback\\": "src/NodejsPhpFallback/"
         }
     },
     "scripts": {
@@ -38,6 +38,9 @@
         ]
     },
     "extra": {
+        "branch-alias": {
+            "dev-master": "2.x-dev"
+        },
         "class": "NodejsPhpFallback\\ComposerPlugin"
     }
 }
