@@ -17,13 +17,18 @@ class ComposerPlugin implements PluginInterface, EventSubscriberInterface
         $this->io = $io;
     }
 
+    public function deactivate(Composer $composer, IOInterface $io)
+    {
+        $this->io = $io;
+    }
+
     public static function getSubscribedEvents()
     {
-        return array(
-            'post-autoload-dump' => array(
-                array('onAutoloadDump', 0),
-            ),
-        );
+        return [
+            'post-autoload-dump' => [
+                ['onAutoloadDump', 0],
+            ],
+        ];
     }
 
     public function onAutoloadDump(Event $event)
@@ -31,13 +36,8 @@ class ComposerPlugin implements PluginInterface, EventSubscriberInterface
         NodejsPhpFallback::install($event);
     }
 
-    public function deactivate(Composer $composer, IOInterface $io)
-    {
-        // Not needed
-    }
-
     public function uninstall(Composer $composer, IOInterface $io)
     {
-        // Not needed
+        $this->io = $io;
     }
 }
