@@ -9,12 +9,12 @@ class UpperCaseWrapperWithNode extends Wrapper
 {
     public function compile()
     {
-        return 'compile:' . $this->getPath('source.upper') . ':' . strtoupper($this->getSource());
+        return 'compile:'.$this->getPath('source.upper').':'.strtoupper($this->getSource());
     }
 
     public function fallback()
     {
-        return 'fallback:' . $this->getPath('source.upper') . ':' . strtoupper($this->getSource());
+        return 'fallback:'.$this->getPath('source.upper').':'.strtoupper($this->getSource());
     }
 }
 
@@ -37,18 +37,18 @@ class NodejsPhpFallbackTest extends TestCase
     {
         $wrapper = new UpperCaseWrapperWithoutNode('foo');
 
-        static::removeDirectory(__DIR__ . '/../node_modules');
-        mkdir(__DIR__ . '/../node_modules/foo', 0777, true);
-        touch(__DIR__ . '/../node_modules/foo/foo');
-        chmod(__DIR__ . '/../node_modules/foo/foo', 0777);
+        static::removeDirectory(__DIR__.'/../node_modules');
+        mkdir(__DIR__.'/../node_modules/foo', 0777, true);
+        touch(__DIR__.'/../node_modules/foo/foo');
+        chmod(__DIR__.'/../node_modules/foo/foo', 0777);
         $withFallback = $wrapper->execModuleScript('foo', 'foo', 'foo', function () {
             return 42;
         });
         $withoutFallback = $wrapper->execModuleScript('foo', 'foo', 'foo');
-        static::removeDirectory(__DIR__ . '/../node_modules');
+        static::removeDirectory(__DIR__.'/../node_modules');
 
         $this->assertSame(42, $withFallback);
-        $this->assertSame(null, $withoutFallback);
+        $this->assertNull($withoutFallback);
     }
 
     public function testGetPath()
@@ -57,8 +57,8 @@ class NodejsPhpFallbackTest extends TestCase
         $this->assertSame(__FILE__, $wrapper->getPath());
 
         $wrapper = new UpperCaseWrapperWithNode('does/not/exists');
-        $this->assertSame(sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'source.tmp', $wrapper->getPath());
-        $this->assertSame(sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'bar', $wrapper->getPath('bar'));
+        $this->assertSame(sys_get_temp_dir().DIRECTORY_SEPARATOR.'source.tmp', $wrapper->getPath());
+        $this->assertSame(sys_get_temp_dir().DIRECTORY_SEPARATOR.'bar', $wrapper->getPath('bar'));
     }
 
     public function testGetSource()
@@ -73,44 +73,44 @@ class NodejsPhpFallbackTest extends TestCase
     public function testGetResult()
     {
         $wrapper = new UpperCaseWrapperWithNode(__FILE__);
-        $this->assertSame('compile:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $wrapper->getResult());
+        $this->assertSame('compile:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $wrapper->getResult());
 
         $wrapper = new UpperCaseWrapperWithoutNode(__FILE__);
-        $this->assertSame('fallback:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $wrapper->getResult());
+        $this->assertSame('fallback:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $wrapper->getResult());
     }
 
     public function testExec()
     {
         $wrapper = new UpperCaseWrapperWithNode(__FILE__);
-        $this->assertSame('compile:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $wrapper->exec());
+        $this->assertSame('compile:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $wrapper->exec());
 
         $wrapper = new UpperCaseWrapperWithoutNode(__FILE__);
-        $this->assertSame('fallback:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $wrapper->exec());
+        $this->assertSame('fallback:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $wrapper->exec());
     }
 
     public function testWrite()
     {
-        $file = sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'test.tmp';
+        $file = sys_get_temp_dir().DIRECTORY_SEPARATOR.'test.tmp';
 
         $wrapper = new UpperCaseWrapperWithNode(__FILE__);
         $wrapper->write($file);
         $text = file_get_contents($file);
         unlink($file);
-        $this->assertSame('compile:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $text);
+        $this->assertSame('compile:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $text);
 
         $wrapper = new UpperCaseWrapperWithoutNode(__FILE__);
         $wrapper->write($file);
         $text = file_get_contents($file);
         unlink($file);
-        $this->assertSame('fallback:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), $text);
+        $this->assertSame('fallback:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), $text);
     }
 
     public function testToString()
     {
         $wrapper = new UpperCaseWrapperWithNode(__FILE__);
-        $this->assertSame('compile:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), strval($wrapper));
+        $this->assertSame('compile:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), strval($wrapper));
 
         $wrapper = new UpperCaseWrapperWithoutNode(__FILE__);
-        $this->assertSame('fallback:' . __FILE__ . ':' . strtoupper(file_get_contents(__FILE__)), strval($wrapper));
+        $this->assertSame('fallback:'.__FILE__.':'.strtoupper(file_get_contents(__FILE__)), strval($wrapper));
     }
 }
