@@ -1,5 +1,5 @@
 <?php
 
-require __DIR__ . '/../../vendor/autoload.php';
-require __DIR__ . '/TestCase.php';
-require __DIR__ . '/CaptureIO.php';
+require __DIR__.'/../../vendor/autoload.php';
+require __DIR__.'/TestCase.php';
+require __DIR__.'/CaptureIO.php';
